requirements.txt
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
openai==0.28.1
python-dotenv==1.0.0
